reset: re-add support the now-deprecated --stdin option
Just like with other Git commands, this option makes it read the paths
from the standard input. It comes in handy when resetting many, many
paths at once and wildcards are not an option (e.g. when the paths are
generated by a tool).

Note: we first parse the entire list and perform the actual reset action
only in a second phase. Not only does this make things simpler, it also
helps performance, as do_diff_cache() traverses the index and the
(sorted) pathspecs in simultaneously to avoid unnecessary lookups.

This feature is marked deprecated because it was superseded by support
for the `--pathspec-from-file=-` option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>